some refactoring and added unit test to windowvalidator
