+++ handlers/new-product-api/src/main/scala/com/gu/newproduct/api/productcatalog/Catalog.scala
@@ -2,8 +2,6 @@ package com.gu.newproduct.api.productcatalog
 
 import java.time.DayOfWeek
 import java.time.DayOfWeek._
-import com.gu.newproduct.api.addsubscription.validation
-import com.gu.newproduct.api.addsubscription.validation._
 
 case class Catalog(
   voucherWeekend: Plan,
@@ -22,7 +20,7 @@ object NewProductApi {
     val mondayRule = DaysOfWeekRule(List(MONDAY))
     val voucherWeekendDateRules = StartDateRules(Some(weekendRule), Some(voucherWindowRule))
     val voucherWeekend = Plan(PlanId("voucher_weekend"), voucherWeekendDateRules)
-    val voucherEveryDayDateRules = validation.StartDateRules(Some(mondayRule), Some(voucherWindowRule))
+    val voucherEveryDayDateRules = StartDateRules(Some(mondayRule), Some(voucherWindowRule))
     val voucherEveryDay = Plan(PlanId("voucher_everyDay"), voucherEveryDayDateRules)
     val monthlyContributionWindow = WindowRule(
       maybeSize = Some(WindowSizeDays(1)),
@@ -40,3 +38,14 @@ object NewProductApi {
 }
 case class PlanId(value: String) extends AnyVal
 case class Plan(id: PlanId, startDateRules: StartDateRules = StartDateRules())
+
+case class DelayDays(value: Int) extends AnyVal
+case class WindowSizeDays(value: Int) extends AnyVal
+
+sealed trait DateRule
+
+case class StartDateRules(daysOfWeekRule: Option[DaysOfWeekRule] = None, windowRule: Option[WindowRule] = None)
+
+case class DaysOfWeekRule(allowedDays: List[DayOfWeek]) extends DateRule
+
+case class WindowRule(maybeCutOffDay: Option[DayOfWeek], maybeStartDelay: Option[DelayDays], maybeSize: Option[WindowSizeDays]) extends DateRule
